Simplify Junctions tests, add indirection test
